modificacion listado propiedades

+++ src/main/java/ar/edu/unlam/tallerweb1/modelo/TipoPropiedad.java
@@ -1,7 +1,5 @@
 package ar.edu.unlam.tallerweb1.modelo;
 
-import net.bytebuddy.dynamic.loading.InjectionClassLoader;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -12,26 +10,27 @@ public class TipoPropiedad {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 
-    private Long id;
+    private Integer id;
     private String descripcion;
 
     public TipoPropiedad(String descripcion) {
         this.descripcion = descripcion;
     }
 
-    public Long getId() {
-        return id;
-    }
+	public Integer getId() {
+		return id;
+	}
 
-    public void setId(Long id) {
-        this.id = id;
-    }
+	public void setId(Integer id) {
+		this.id = id;
+	}
 
-    public String getDescripcion() {
-        return descripcion;
-    }
+	public String getDescripcion() {
+		return descripcion;
+	}
 
-    public void setDescripcion(String descripcion) {
-        this.descripcion = descripcion;
-    }
+	public void setDescripcion(String descripcion) {
+		this.descripcion = descripcion;
+	}
+
 }
